cli: moved bulk of create implementation over to lib
