+++ packages/cli/src/lib/create/factories/frontendPlugin.ts
